v1.0.1
- Small updates to game manager singleton and cocos2d

+++ Source/libs/cocos2d-iphone/cocos2d-ui/CCBReader/CCBReader.m
@@ -172,7 +172,8 @@ static inline unsigned int readVariableLengthIntFromArray(const uint8_t* buffer,
     
 done:
     *value = result;
-    return ptr - buffer;
+    //return ptr - buffer;
+    return (unsigned int)(ptr - buffer);
 }
